4_signals: introduce signals lecture examples
This semester is a first time when that lecture is being given
'in a full size'. Signal examples mostly show how to work with
signal(), sigaction(), some typical errors. Also, the final and
the biggest example (12) consists of three files: two for a
coroutine library written on signal stacks and long jumps, and
one with its usage.